+++ src/components/Editor/Forms/About/Operations/DeleteForm.js
@@ -1,9 +1,8 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useContext } from 'react';
 import { DataContext } from '../../../../../DataContext';
-import TextFieldCtrl from '../../Controllers/TextFieldCtrl';
+import FormInputs from './FormInputs';
 import ButtonCtrl from '../../Controllers/ButtonCtrl';
 import { useForm } from 'react-hook-form';
-import { deleteController } from '../../../../../firestore';
 import { makeStyles } from '@material-ui/core/styles';
 import styled from 'styled-components';
 
@@ -49,52 +48,16 @@ const DeleteForm = ({ id, university, major, course, startYear, endYear }) => {
   return (
     <form className={classes.form} onSubmit={handleSubmit(onSubmit)}>
       <h1>EDUCATION</h1>
-      <TextFieldCtrl
-        name={'id'}
-        label={'Unique Identifier'}
-        value={id}
+      <FormInputs
+        id={id}
+        university={university}
+        major={major}
+        course={course}
+        startYear={startYear}
+        endYear={endYear}
         control={control}
         readOnly={true}
       />
-      <FlexContainer>
-        <TextFieldCtrl
-          name={'startYear'}
-          label={'Start Year'}
-          value={startYear}
-          control={control}
-          readOnly={true}
-        />
-        <TextFieldCtrl
-          name={'endYear'}
-          label={'End Year'}
-          value={endYear}
-          control={control}
-          readOnly={true}
-        />
-      </FlexContainer>
-      <FlexContainer>
-        <TextFieldCtrl
-          name={'university'}
-          label={'University'}
-          value={university}
-          control={control}
-          readOnly={true}
-        />
-        <TextFieldCtrl
-          name={'major'}
-          label={'Major'}
-          value={major}
-          control={control}
-          readOnly={true}
-        />
-        <TextFieldCtrl
-          name={'course'}
-          label={'Course'}
-          value={course}
-          control={control}
-          readOnly={true}
-        />
-      </FlexContainer>
       <Buttons>
         <ButtonCtrl title={'Delete'} />
       </Buttons>
@@ -105,9 +68,5 @@ const DeleteForm = ({ id, university, major, course, startYear, endYear }) => {
 export default DeleteForm;
 
 const Buttons = styled.div`
-  display: flex;
-`;
-
-const FlexContainer = styled.div`
   display: flex;
 `;
