+++ src/components/Editor/Forms/About/Sections/Education.js
@@ -1,73 +1,13 @@
-import React, { useContext } from 'react';
-import { DataContext } from '../../../../../DataContext';
-import TextAreaCtrl from '../../Controllers/TextFieldCtrl';
-import ButtonCtrl from '../../Controllers/ButtonCtrl';
-import { useForm } from 'react-hook-form';
-import { createController } from '../../../../../firestore';
-import { makeStyles } from '@material-ui/core/styles';
+import React from 'react';
 import styled from 'styled-components';
+import CreatForm from '../Operations/CreateForm';
 import UpdateForm from '../Operations/UpdateForm';
 import DeleteForm from '../Operations/DeleteForm';
 
-const useStyles = makeStyles({
-  root: {
-    flexGrow: 1,
-    background: 'transparent',
-  },
-  tabs: {
-    overflow: 'hidden',
-    background: '#EEFBFB',
-  },
-  form: {
-    display: 'flex',
-    flexDirection: 'column',
-    padding: '1rem',
-    width: '100%',
-  },
-});
-
-/**
- * useForm requires a default state/values to avoid the following error:
- * 'A component is changing an uncontrolled input to be controlled.'
- */
-const defaultValues = {
-  course: '',
-  major: '',
-  university: '',
-  startYear: '',
-  endYear: '',
-};
-
 const Education = ({ operation, education }) => {
-  const classes = useStyles();
-
-  const { onCreate } = useContext(DataContext);
-
-  const { handleSubmit, reset, control } = useForm({ defaultValues });
-
-  const onSubmit = async (data) => {
-    onCreate({ data, collection: 'education' });
-  };
-
   return (
     <Container>
-      {operation === 0 && (
-        <form className={classes.form} onSubmit={handleSubmit(onSubmit)}>
-          <h1>EDUCATION</h1>
-          <FlexContainer>
-            <TextAreaCtrl name={'university'} label={'University'} control={control} />
-            <TextAreaCtrl name={'major'} label={'Major'} control={control} />
-            <TextAreaCtrl name={'course'} label={'Course'} control={control} />
-          </FlexContainer>
-          <FlexContainer>
-            <TextAreaCtrl name={'startYear'} label={'Start Year'} control={control} />
-            <TextAreaCtrl name={'endYear'} label={'End Year'} control={control} />
-          </FlexContainer>
-          <Buttons>
-            <ButtonCtrl reset={reset} initialState={defaultValues} />
-          </Buttons>
-        </form>
-      )}
+      {operation === 0 && <CreatForm />}
       {operation === 1 &&
         education?.map(({ id, university, major, course, startYear, endYear }) => (
           <UpdateForm
@@ -111,12 +51,4 @@ const Container = styled.div`
     color: ${({ theme }) => theme.primaryDark};
     margin: 1rem 0rem 1rem 0.5rem; // top right bottom left
   }
-`;
-
-const Buttons = styled.div`
-  display: flex;
-`;
-
-const FlexContainer = styled.div`
-  display: flex;
 `;
